requirements.txt
openai
youtube-transcript-api
moviepy
openai-whisper
googletrans==4.0.0-rc1
stramlit